chore: improve ThemeProvider exports
